zero/meta-arducam: Add arducam device-tree
This commit adds a dts overlay file to enable recognition of the
Arducam camera.

The DTS and driver for the Arducam camera have been merged into
the raspberrypi linux kernel as of version 5.15.38

However, the kernel used in kirkstone is version 5.15.34, so we
will apply the dts file merged in the following commit as a patch.

  - repository: https://github.com/raspberrypi/linux.git
    hash :1f8131e:
    subject: SQUASH: overlays: arducam-pivariety: Strip leading zero

Signed-off-by: KantaTamura <tkanta496@gmail.com>
Co-authored-by:: Takuya Sasaki <takuya.sasaki@spacecubics.com>
Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>