zero: csp-handler: Add file work thread
CSP commands related to file operation may take time to process.
Therefore, a dedicated thread is provided to handle CSP packets
sequentially via a glib-2.0 queue. This commit only adds the
framework for this functionality.

Additionally, the CSP port for file-related commands is set to `13`
to maintain consistency with MAIN and ADCS.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>